chore: :truck: renamed files to global vs local
